% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict via Linear Regression Model}
\usage{
predict(model, newdata)
}
\arguments{
\item{model}{describle model formula}

\item{newdata}{build from data.frame}
}
\value{
return a list consist of coefficiences and formula
}
\description{
train a linear model
}
