% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lm}
\alias{print.lm}
\title{Linear Regression Model}
\usage{
\method{print}{lm}(model)
}
\arguments{
\item{model}{the model you want to display}
}
\value{
return a list consist of coefficiences and formula
}
\description{
train a linear model
}
