% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_reg.R
\name{save_model}
\alias{save_model}
\title{Linear Regression Model}
\usage{
save_model(model, filename)
}
\arguments{
\item{filename}{the save model file name}

\item{mode}{the linear model you want to save}
}
\value{
none
}
\description{
save model
}
