% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_reg.R
\name{lin_reg}
\alias{lin_reg}
\title{Linear Regression Model}
\usage{
lin_reg(formula, data)
}
\arguments{
\item{formula}{describle model formula}

\item{data}{data build from data.frame}
}
\value{
return a list consist of coefficiences and formula
}
\description{
train a linear model
}
